Use event-based flow to delete unreferenced files
Previously, to delete a file from the file store, application code would call
`FileService.deleteFile` and pass it a function to remove the reference to the
file from the appropriate child tables.

That worked well enough for simple cases where there was a 1:1 relationship
between files and references to them. But we're starting to see more complex
1:N relationships, and in those cases, it's not always clear which code should
be responsible for calling `FileService.deleteFile`.

Improve the situation by moving to an event-based scheme. When application code
deletes a reference to a file, instead of doing

    fileService.deleteFile(fileId) {
        deleteRowsThatReferToFile(fileId)
    }

it should now do

    deleteRowsThatReferToFile(fileId)
    eventPublisher.publishEvent(FileReferenceDeletedEvent(fileId))

`FileService` listens for the event and takes care of deleting the file from
the file store if there are no other references to the file.

With this setup, in a 1:N situation, the event can be published as each reference
is deleted, without regard to whether or not there are other references. When the
last one is gone, the file will be deleted.

Since the files in question may be video files that are associated with Mux
assets, and we only want to delete those when the video is no longer
referenced, Mux asset deletion is now handled the same way as deletion of
thumbnail images: it's triggered by the `FileDeletionStartedEvent` that's
published by `FileService` before the file is actually removed from the file
store, rather than by a video-specific event from `ActivityMediaService`. That
way as we add video support to more places in the app, we won't have to
remember to publish the correct deletion event everywhere.

We use a "scan for references" approach rather than maintaining a reference count
because it's simpler to make robust: it's impossible to fail to increment or
decrement the reference count. We also have a relatively low rate of file
deletion, so scanning a bunch of tables for references should be less expensive
overall than keeping reference counts up to date.

+++ src/main/kotlin/com/terraformation/backend/file/mux/MuxService.kt
@@ -11,7 +11,7 @@ import com.terraformation.backend.file.FileService
 import com.terraformation.backend.file.JpegConverter
 import com.terraformation.backend.file.ThumbnailNotReadyException
 import com.terraformation.backend.file.VideoStreamNotFoundException
-import com.terraformation.backend.file.event.VideoFileDeletedEvent
+import com.terraformation.backend.file.event.FileDeletionStartedEvent
 import com.terraformation.backend.file.event.VideoFileUploadedEvent
 import com.terraformation.backend.log.perClassLogger
 import io.jsonwebtoken.Jwts
@@ -282,15 +282,17 @@ class MuxService(
   }
 
   @EventListener
-  fun on(event: VideoFileDeletedEvent) {
-    val assetId =
-        dslContext.fetchValue(MUX_ASSETS.ASSET_ID, MUX_ASSETS.FILE_ID.eq(event.fileId)) ?: return
+  fun on(event: FileDeletionStartedEvent) {
+    if (canConvertToJpeg(event.contentType)) {
+      val assetId =
+          dslContext.fetchValue(MUX_ASSETS.ASSET_ID, MUX_ASSETS.FILE_ID.eq(event.fileId)) ?: return
 
-    // Asynchronously delete the asset from Mux; we don't want the event listener to block if
-    // Mux takes a long time to respond.
-    jobScheduler.enqueue<MuxService> { deleteMuxAssetIfExists(assetId) }
+      // Asynchronously delete the asset from Mux; we don't want the event listener to block if
+      // Mux takes a long time to respond.
+      jobScheduler.enqueue<MuxService> { deleteMuxAssetIfExists(assetId) }
 
-    dslContext.deleteFrom(MUX_ASSETS).where(MUX_ASSETS.FILE_ID.eq(event.fileId)).execute()
+      dslContext.deleteFrom(MUX_ASSETS).where(MUX_ASSETS.FILE_ID.eq(event.fileId)).execute()
+    }
   }
 
   @Suppress("MemberVisibilityCanBePrivate") // Called by JobRunr
